+++ test/test_reg.py
@@ -1,9 +1,10 @@
 import pytest
 import random
 from pyv.reg import * 
+from pyv.clocked import RegBase
 
 def test_reg():
-    RegBase._clearRegList()
+    RegBase.clear()
 
     reg = Reg()
     RegBase.reset()
@@ -12,15 +13,13 @@ def test_reg():
     reg.next.write(0x42)
     assert reg.cur.read() == 0
 
-    reg._prepareNextVal()
-    reg._tick()
+    RegBase.tick()
     assert reg.cur.read() == 0x42
 
     reg.next.write(0x69)
     assert reg.cur.read() == 0x42
 
-    reg._prepareNextVal()
-    reg._tick()
+    RegBase.tick()
     assert reg.cur.read() == 0x69
 
 def test_regbase():
@@ -35,7 +34,7 @@ def test_regbase():
         reg._tick()
 
 def test_regX():
-    RegBase._clearRegList()
+    RegBase.clear()
 
     reg = RegX('A', 'B')
     RegBase.reset()
@@ -43,8 +42,7 @@ def test_regX():
     reg.next.write('A', 42, 'B', 69)
     assert reg.cur._val['A']._val == 0
     assert reg.cur._val['B']._val == 0
-    reg._prepareNextVal()
-    reg._tick()
+    RegBase.tick()
     assert reg.cur._val['A']._val == 42
     assert reg.cur._val['B']._val == 69
 
@@ -54,7 +52,7 @@ def test_regX():
 
 
 def test_regfile():
-    RegBase._clearRegList()
+    RegBase.clear()
 
     rf = Regfile()
 
@@ -65,18 +63,32 @@ def test_regfile():
     assert val2 == 0
 
     # Write some values
-    rf.write(14, 0xdeadbeef)
+    rf.writeRequest(14, 0xdeadbeef)
+    assert rf.regs[14] == 0
+    rf._tick()
     assert rf.regs[14] == 0xdeadbeef
 
-    rf.write(2, 0x42)
+    rf.writeRequest(2, 0x42)
+    assert rf.regs[2] == 0
+    rf._tick()
     assert rf.regs[2] == 0x42
 
+    rf._tick()
+    assert rf.regs[2] == 0x42
+    assert rf.regs[14] == 0xdeadbeef
+
     # Write to x0
-    rf.write(0, 0xdeadbeef)
+    rf.writeRequest(0, 0xdeadbeef)
+    assert rf.regs[0] == 0
+    rf._tick()
     assert rf.regs[0] == 0
 
+    # Test reset
+    rf._reset()
+    assert rf.regs == [0 for _ in range(0,32)]
+
 def test_regChain():
-    RegBase._clearRegList()
+    RegBase.clear()
 
     A = Reg()
     B = Reg()
@@ -90,7 +102,7 @@ def test_regChain():
 
     A.next.write(0x42)
 
-    RegBase._updateRegs()
+    RegBase.tick()
     assert A.cur.read() == 0x42
     assert B.cur.read() == 0
     assert C.cur.read() == 0
@@ -98,26 +110,26 @@ def test_regChain():
 
     A.next.write(0)
 
-    RegBase._updateRegs()
+    RegBase.tick()
     assert A.cur.read() == 0
     assert B.cur.read() == 0x42
     assert C.cur.read() == 0
     assert D.cur.read() == 0
 
-    RegBase._updateRegs()
+    RegBase.tick()
     assert A.cur.read() == 0
     assert B.cur.read() == 0
     assert C.cur.read() == 0x42
     assert D.cur.read() == 0
 
-    RegBase._updateRegs()
+    RegBase.tick()
     assert A.cur.read() == 0
     assert B.cur.read() == 0
     assert C.cur.read() == 0
     assert D.cur.read() == 0x42
 
 def test_regChainX():
-    RegBase._clearRegList()
+    RegBase.clear()
 
     A = RegX('A', 'B')
     B = RegX('A', 'B')
@@ -131,7 +143,7 @@ def test_regChainX():
 
     A.next.write('A',45,'B',78)
 
-    RegBase._updateRegs()
+    RegBase.tick()
     assert A.cur.read() == {'A':45, 'B':78}
     assert B.cur.read() == {'A':0, 'B':0}
     assert C.cur.read() == {'A':0, 'B':0}
@@ -139,26 +151,26 @@ def test_regChainX():
 
     A.next.write('A',0,'B',0)
 
-    RegBase._updateRegs()
+    RegBase.tick()
     assert A.cur.read() == {'A':0, 'B':0}
     assert B.cur.read() == {'A':45, 'B':78}
     assert C.cur.read() == {'A':0, 'B':0}
     assert D.cur.read() == {'A':0, 'B':0}
 
-    RegBase._updateRegs()
+    RegBase.tick()
     assert A.cur.read() == {'A':0, 'B':0}
     assert B.cur.read() == {'A':0, 'B':0}
     assert C.cur.read() == {'A':45, 'B':78}
     assert D.cur.read() == {'A':0, 'B':0}
 
-    RegBase._updateRegs()
+    RegBase.tick()
     assert A.cur.read() == {'A':0, 'B':0}
     assert B.cur.read() == {'A':0, 'B':0}
     assert C.cur.read() == {'A':0, 'B':0}
     assert D.cur.read() == {'A':45, 'B':78}
 
 def test_shiftReg():
-    RegBase._clearRegList()
+    RegBase.clear()
 
     depth = 32
     A = ShiftReg(depth)
@@ -172,7 +184,7 @@ def test_shiftReg():
     # Fill shift register
     for i in range(depth):
         A.serIn.write(inputs[i])
-        RegBase._updateRegs()
+        RegBase.tick()
         #print("i = {}, regs = {}".format(i, A.regs))
 
     # Drain shift register
@@ -180,15 +192,15 @@ def test_shiftReg():
     for i in range(depth):
         #print("i = {}, regs = {}".format(i, A.regs))
         assert A.serOut.read() == inputs[i]
-        RegBase._updateRegs()
+        RegBase.tick()
     
     # Test warning when input wider than 1 bit
     A.serIn.write(4)
     with pytest.warns(UserWarning):
-        RegBase._updateRegs()
+        RegBase.tick()
 
 def test_shiftRegParallel():
-    RegBase._clearRegList()
+    RegBase.clear()
 
     depth = 8
     A = ShiftRegParallel(depth)
@@ -199,7 +211,7 @@ def test_shiftRegParallel():
     # Test parallel load
     A.parEnable.write(1)
     A.parIn.write(0xAF)
-    RegBase._updateRegs()
+    RegBase.tick()
     assert A.regs == [1,0,1,0,1,1,1,1]
 
     # Test parallel read
@@ -209,11 +221,11 @@ def test_shiftRegParallel():
     A.parEnable.write(0)
 
     A.serIn.write(1)
-    RegBase._updateRegs()
+    RegBase.tick()
     A.serIn.write(1)
-    RegBase._updateRegs()
+    RegBase.tick()
     A.serIn.write(0)
-    RegBase._updateRegs()
+    RegBase.tick()
 
     # Test parallel read again
     print(A.regs)
@@ -240,20 +252,20 @@ def test_shiftRegParallel():
         elif i == 7:
             assert A.serOut.read() == 0
         
-        RegBase._updateRegs()
+        RegBase.tick()
 
     # Test parallel load with value wider than depth
     # This should issue a warning.
     A.parEnable.write(1)
     A.parIn.write(0xAEADBEEF)
     with pytest.warns(UserWarning):
-        RegBase._updateRegs()
+        RegBase.tick()
 
     # Test parallel load with value narrower than depth
     # This should make sure that the length of the list remains the same.
     A.parEnable.write(1)
     A.parIn.write(0x47)
-    RegBase._updateRegs()
+    RegBase.tick()
     assert len(A.regs) == depth
 
 def test_reset():