+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.2.0
+- **New**: Implemented abstract `Clocked` class to unite handling of registers and memories
+    - Write operations on memories and register files will _not_ commit until the next clock tick
+      anymore
+        - This is to ensure data coherency throughout the current simulation cycle 
+
 # 0.1.0
 
 - Implemented a whole new simulation model